1. Update the command lines `toolbox-iop-duplicate-id` 2. Add function `find_duplicate_ids` , `load_json_line` 3. Update README
