+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "toolbox-iop"
-version = "0.1.3"
+version = "0.1.4"
 description = "The toolbox for json load save and print, you may find more function in the future."
 readme = "README.md"
 authors = [
@@ -11,10 +11,11 @@ dependencies = [
     "rich>=10.11.0",
     "tqdm>=4.64.1",
 ]
+license = { text = "MIT" }
 
 [project.scripts]
 toolbox-iop-print = "toolbox_iop.cli:print_main"
-toolbox-iop-check-jid = "toolbox_iop.cli:check_jid"
+toolbox-iop-duplicate = "toolbox_iop.cli:check_main"
 
 [build-system]
 requires = ["hatchling"]
